Course ownership

+++ controllers/courses.controller.js
@@ -38,8 +38,14 @@ exports.getCourse = asyncHandler(async(req,res,next)=>{
 //@access private
 exports.createCourse = asyncHandler(async(req,res,next)=>{
   req.body.bootcamp = req.params.bootcampId
+  req.body.user= req.user.id;
 
-  let bootcamp = await Bootcamp.findById(req.params.bootcampId)
+  let bootcamp = await Bootcamp.findById(req.params.bootcampId);
+
+   //Make sure user is bootcamp owner
+   if (bootcamp.user.toString()!==req.user.id && req.user.role !== 'admin') {
+    return next(new ErrorResponse(`User ${req.user.id} is not authorized to add a course to this bootcamp`,401))
+    }
   
   if (!bootcamp) {
     return next(new ErrorResponse(`No bootcamp found with the id ${req.params.id}`))
@@ -57,10 +63,14 @@ exports.updateCourse = asyncHandler(async(req,res,next)=>{
   
   const course = await Course.findById(req.params.id)
 
-  
   if (!course) {
     return next(new ErrorResponse(`No course found with the id ${req.params.id}`))
   }
+
+   //Make sure user is courser owner
+   if (course.user.toString()!==req.user.id && req.user.role !== 'admin') {
+    return next(new ErrorResponse(`User ${req.user.id} is not authorized to update this course`,401))
+    }
 
   course = await Course.findByIdAndUpdate(req.params.id,req.body,{
     new:true,
@@ -81,6 +91,10 @@ exports.deleteCourse = asyncHandler(async(req,res,next)=>{
   if (!course) {
     return next(new ErrorResponse(`No course found with the id ${req.params.id}`))
   }
+
+  if (course.user.toString()!==req.user.id && req.user.role !== 'admin') {
+    return next(new ErrorResponse(`User ${req.user.id} is not authorized to delete this course`,401))
+    }
 
   await Course.deleteOne();
   res.status(200).json({success:true,data:{}})
